app/requirements.txt
functions-framework==3.*
spotipy==2.22.0
google-cloud-secret-manager==2.13.0
